backend/pyproject.toml
[project]
name = "ice-cold-beer-scorekeeper"
version = "0.1.0"
description = "Backend for Ice Cold Beer Scorekeeper app"
authors = [
    {name = "copilot-swe-agent[bot]",email = "198982749+Copilot@users.noreply.github.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
